pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-mc-tiktok"
version = "1.0.0"
description = "Automate creating TikTok-style videos from Reddit stories with Minecraft parkour background"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["reddit", "tiktok", "video", "automation", "tts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "praw>=7.7.1",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pyttsx3>=2.90",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "httpx>=0.27.0",
    "click>=8.1.7",
    "ffmpeg-python>=0.2.0",
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
reddit-mc-tiktok = "reddit_mc_tiktok.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/reddit-mc-tiktok"
Repository = "https://github.com/yourusername/reddit-mc-tiktok.git"
Issues = "https://github.com/yourusername/reddit-mc-tiktok/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
